refactor: isolate geometry and mesh queue
