+++ docs/superpowers/plans/2026-08-17-modularization-plan.md
@@ -76,8 +76,9 @@
 **Files:** `lib/ChunkMesher.lua`, new `lib/GeometryBuilder.lua`, new `lib/MeshQueue.lua`, new `lib/MeshRuntime.lua`, worker/thread tests
 
 - [x] Characterize stream counts, fingerprints, job ordering, budget slicing, worker messages, and runtime mesh eviction.
-- [ ] Move map-to-stream construction into `GeometryBuilder` with no engine or queue dependencies.
-- [ ] Move sliced scheduling and job state into `MeshQueue`.
+- [x] Move map-to-stream construction into `GeometryBuilder` with no GPU or queue dependencies.
+- [x] Move queue state, deduplication, completion, and cancellation into `MeshQueue`.
+- [ ] Move budget-slice orchestration and worker dispatch policy into `MeshQueue`.
 - [x] Move GPU mesh upload, cache-entry release/swap rules, and live-set eviction into `MeshRuntime`.
 - [ ] Keep the worker protocol and `ChunkMesher` public calls stable; verify serial/threaded equivalence.
