+++ lib/ChunkMesher.lua
@@ -52,12 +52,12 @@ local V = ...
 
 local Assets = require("src.render.Assets")
 local Structures = V.require("Structures")
-local TileShape = V.require("TileShape")
 local Voxel3D = V.require("Voxel3D")
 local Budget = V.require("BuildBudget")
 local MeshCache = V.require("MeshCache")
-local GridKey = V.require("GridKey")
 local MeshRuntime = V.require("MeshRuntime")
+local MeshQueue = V.require("MeshQueue")
+local GeometryBuilder = V.require("GeometryBuilder")
 
 -- ffi is gone (sandbox) and this engine's love.data ByteData carries no
 -- accessors, so the native float buffers are gone with them: the table
@@ -65,44 +65,10 @@ local MeshRuntime = V.require("MeshRuntime")
 
 local ChunkMesher = {}
 local Runtime = MeshRuntime.new()
-
--- Ring of border blocks meshed around the body, matching the width
--- TileRenderer draws so the two modes end at the same place.
-local RING = 3
-
--- A sliver of a texel, to keep a quad's sampling inside its own tile.
--- Without any inset the perspective rasteriser lands on a NEIGHBOURING
--- tile's texel along the shared edge and stitches bright seams across the
--- whole map.
---
--- It has to be a sliver and not, as it first was, half a texel. A tile is
--- 8 texels of art across 8 world pixels -- one texel per pixel exactly --
--- and insetting the uv by half a texel at each end squeezes that art into
--- a 7-texel sample range while the quad still covers 8 world pixels. The
--- art then advances 7/8 of a texel per pixel: boundaries drift off the
--- pixel grid, one art pixel gets sampled twice and another never at all.
--- Nothing showed it until the voxel wireframe drew the grid those pixels
--- were supposed to be sitting on. Interpolation error is nowhere near a
--- fiftieth of a texel, so this is as safe against bleed and costs 0.25% of
--- a pixel of drift across a whole tile.
-local INSET = 0.02
-
--- The south face of a volume is the artwork itself, so it draws at full
--- brightness; its top face darkens a touch so the plateau behind a
--- standing drawing reads as depth rather than repeating the same art at
--- the same energy.
-local VOLUME_TOP_SHADE = 0.85
+local Queue = MeshQueue.new()
 
 local cache = {}     -- map id -> { full = mesh|false, body = ..., grass = ... }
 local gen = {}       -- map id -> generation, bumped by invalidate/evict
-
--- Horizontal neighbours: tile step, face direction id (see Voxel3D).
-local SIDES = {
-  { 1, 0, 1 },    -- +X east
-  { -1, 0, 2 },   -- -X west
-  { 0, 1, 5 },    -- +Z south
-  { 0, -1, 6 },   -- -Z north
-}
 
 -- ------------------------------------------------------------ vertex sinks
 
@@ -211,616 +177,7 @@ end
 -- faces around them still belong to the GROUND that exposes them.
 --
 -- Omitted, water stays in the terrain mesh exactly as it always did, which
--- is what the headless geometry() below and the sun's own pass both want.
-local function runGeometry(map, bodyOnly, masks, sink, waterSink)
-  local push = sink.push
-  local waterPush = waterSink and waterSink.push or nil
-  local tileset = map.tileset
-  local S = Structures.forMap(map)
-  local perRow = tileset.tilesPerRow or 16
-  local atlasW = tileset.imageWidth or (perRow * 8)
-  local atlasH = tileset.imageHeight or 48
-
-  local function heightAt(tx, ty)
-    local k = GridKey.of(tx, ty)
-    if S.skip[k] then return 0 end
-    local run = S.runs[k]
-    if run then return run.h end
-    local s = S.shapeAt[k]
-    return s and s.h or 0
-  end
-
-  -- one atlas-rect UV, optionally cropped to art rows [vTop, vBot] of 8
-  local function uvRect(tile, vTop, vBot)
-    local ax = (tile % perRow) * 8
-    local ay = math.floor(tile / perRow) * 8
-    local vi = math.min(INSET, (vBot - vTop) / 4)
-    return (ax + INSET) / atlasW, (ax + 8 - INSET) / atlasW,
-           (ay + vTop + vi) / atlasH, (ay + vBot - vi) / atlasH
-  end
-
-  -- ------------------------------------------------------ ambient occlusion
-  --
-  -- Ambient light is what reaches a surface from the sky at large, so it is
-  -- blocked by how much geometry crowds a point rather than by where the
-  -- sun happens to be -- which makes it the exact complement of the shadow
-  -- pass, and the reason both are worth having. The shadow map draws the
-  -- long directional shadow a building throws; this draws the dark seam in
-  -- every corner the sky cannot see into, at every scale finer than a
-  -- shadow map texel.
-  --
-  -- Baked per vertex, the classic voxel way: each corner counts the
-  -- neighbours that crowd it and steps down once per neighbour, and the
-  -- rasteriser interpolates the steps into a smooth falloff. Costs exactly
-  -- nothing at draw time, and it is resolution-independent -- a screen
-  -- space pass would blur across the pixel grid this whole mode is built
-  -- to keep crisp.
-  --
-  -- (What was here before was a one-directional contact shadow keyed to a
-  -- sun in the northwest: two neighbours, one corner, top faces only.)
-
-  -- Intensity. Both terms below are DARKENING amounts rather than
-  -- multipliers, so this one number scales the whole effect: 1.0 is the
-  -- barely-there first cut, and everything is expressed against it.
-  local AO_STRENGTH = 2.4
-  local AO_STEP = 0.09 * AO_STRENGTH      -- per crowding neighbour, max 3
-  local AO_EDGE = 1 - 0.14 * AO_STRENGTH  -- creases / corners on a face
-  local AO_GROUND = 0.12 * AO_STRENGTH    -- a prop's contact with the floor
-  local AO_RISE = 6                       -- px over which the floor lets go
-  local AO_FLOOR = 0.25                   -- never let a vertex reach black
-
-  -- Both sinks copy a per-corner shade straight out into the vertex stream
-  -- and keep no reference, so these two scratch rows are reused for every
-  -- quad on the map rather than allocating a table per face -- a route
-  -- builds a few hundred thousand of them.
-  local aoTop = { 0, 0, 0, 0 }
-  local aoSide = { 0, 0, 0, 0 }
-
-  -- A top face's four corners, each occluded by the three cells that touch
-  -- it: two edge neighbours and the diagonal between them.
-  local function aoShades(tx, ty, h, shade)
-    local n = heightAt(tx, ty - 1) > h
-    local s = heightAt(tx, ty + 1) > h
-    local e = heightAt(tx + 1, ty) > h
-    local w = heightAt(tx - 1, ty) > h
-    local nw = heightAt(tx - 1, ty - 1) > h
-    local ne = heightAt(tx + 1, ty - 1) > h
-    local sw = heightAt(tx - 1, ty + 1) > h
-    local se = heightAt(tx + 1, ty + 1) > h
-    if not (n or s or e or w or nw or ne or sw or se) then return shade end
-    local function corner(a, b, d)
-      local k = 0
-      if a then k = k + 1 end
-      if b then k = k + 1 end
-      -- a diagonal wedged behind both of its edges adds nothing: the
-      -- corner is already as enclosed as it can get, and counting it
-      -- again is what turns an ordinary inside corner black
-      if d and not (a and b) then k = k + 1 end
-      -- floored, so cranking AO_STRENGTH deepens the seams instead of
-      -- punching holes of pure black through the world
-      return shade * math.max(AO_FLOOR, 1 - AO_STEP * k)
-    end
-    -- corners in topQuad order: NW, NE, SE, SW
-    aoTop[1], aoTop[2] = corner(n, w, nw), corner(n, e, ne)
-    aoTop[3], aoTop[4] = corner(s, e, se), corner(s, w, sw)
-    return aoTop
-  end
-
-  -- The same idea on an upright face, where the crowding is of two kinds:
-  -- the CREASE it rises out of (the band sitting on the ground, or on
-  -- whatever lower neighbour exposed the face) and the INSIDE CORNERS
-  -- where the columns flanking it stand proud of the band. `hl`/`hr` are
-  -- those flanking heights in FACE order -- left then right as seen from
-  -- outside, per LATERAL below -- so the shades line up with sideQuad's
-  -- corners without the caller thinking about compass directions.
-  local LATERAL = {
-    [1] = { 0, 1, 0, -1 },    -- east face:  left south, right north
-    [2] = { 0, -1, 0, 1 },    -- west face:  left north, right south
-    [5] = { -1, 0, 1, 0 },    -- south face: left west,  right east
-    [6] = { 1, 0, -1, 0 },    -- north face: left east,  right west
-  }
-  -- Ground contact for the prebuilt prop quads -- the per-pixel plants,
-  -- signs and lone trees, and the round-tree stamps. Those arrive from
-  -- Structures already finished, so the neighbour counting above has no
-  -- columns to count. What it CAN say is that the ground plane itself
-  -- blocks half the sky, so the closer a voxel sits to it the less ambient
-  -- light reaches it -- which is what plants a prop on the floor instead
-  -- of leaving it looking pasted over the top.
-  local aoProp = { 0, 0, 0, 0 }
-  local function groundShades(c, shade)
-    if type(shade) == "table" then return shade end
-    local y1, y2, y3, y4 = c[1][2], c[2][2], c[3][2], c[4][2]
-    if math.min(y1, y2, y3, y4) >= AO_RISE then return shade end
-    for i = 1, 4 do
-      local t = c[i][2] / AO_RISE
-      aoProp[i] = shade * (t >= 1 and 1 or (1 - AO_GROUND * (1 - t)))
-    end
-    return aoProp
-  end
-
-  local AO_CORNER = math.max(AO_FLOOR, AO_EDGE * AO_EDGE)  -- crease AND flank
-  local function sideShades(hl, hr, y0, y1, crease, shade)
-    if not (crease or hl > y0 or hr > y0) then return shade end
-    -- corners run bottom-left, bottom-right, top-right, top-left
-    local base = crease and AO_EDGE or 1
-    aoSide[1] = shade * (hl > y0 and (crease and AO_CORNER or AO_EDGE) or base)
-    aoSide[2] = shade * (hr > y0 and (crease and AO_CORNER or AO_EDGE) or base)
-    aoSide[3] = shade * (hr > y1 and AO_EDGE or 1)
-    aoSide[4] = shade * (hl > y1 and AO_EDGE or 1)
-    return aoSide
-  end
-
-  local scratchC = { {0,0,0}, {0,0,0}, {0,0,0}, {0,0,0} }
-  local scratchUv = { {0,0}, {0,0}, {0,0}, {0,0} }
-
-  -- `to` routes the quad somewhere other than the main sink -- the water
-  -- surface is the only caller that ever does (see runGeometry's header).
-  local function topQuad(x0, z0, h, tile, shade, to)
-    local u0, u1, v0, v1 = uvRect(tile, 0, 8)
-    local c = scratchC
-    c[1][1], c[1][2], c[1][3] = x0, h, z0
-    c[2][1], c[2][2], c[2][3] = x0 + 8, h, z0
-    c[3][1], c[3][2], c[3][3] = x0 + 8, h, z0 + 8
-    c[4][1], c[4][2], c[4][3] = x0, h, z0 + 8
-
-    local uv = scratchUv
-    uv[1][1], uv[1][2] = u0, v0
-    uv[2][1], uv[2][2] = u1, v0
-    uv[3][1], uv[3][2] = u1, v1
-    uv[4][1], uv[4][2] = u0, v1
-
-    ;(to or push)(c, uv, aoShades(x0 / 8, z0 / 8, h, shade))
-  end
-
-  -- vertical quad for face direction `d` of the tile column at (x0, z0),
-  -- spanning heights [y0, y1] and showing art rows [vTop, vBot] of `tile`.
-  -- Corners run bottom-left, bottom-right, top-right, top-left as seen
-  -- from outside; u follows +X on the north/south faces so a door or sign
-  -- never draws mirrored.
-  local function sideQuad(d, x0, z0, y0, y1, tile, vTop, vBot, shade)
-    local x1, z1 = x0 + 8, z0 + 8
-    local c = scratchC
-    if d == 5 then                                       -- south, at z1
-      c[1][1], c[1][2], c[1][3] = x0, y0, z1
-      c[2][1], c[2][2], c[2][3] = x1, y0, z1
-      c[3][1], c[3][2], c[3][3] = x1, y1, z1
-      c[4][1], c[4][2], c[4][3] = x0, y1, z1
-    elseif d == 6 then                                   -- north, at z0
-      c[1][1], c[1][2], c[1][3] = x1, y0, z0
-      c[2][1], c[2][2], c[2][3] = x0, y0, z0
-      c[3][1], c[3][2], c[3][3] = x0, y1, z0
-      c[4][1], c[4][2], c[4][3] = x1, y1, z0
-    elseif d == 1 then                                   -- east, at x1
-      c[1][1], c[1][2], c[1][3] = x1, y0, z1
-      c[2][1], c[2][2], c[2][3] = x1, y0, z0
-      c[3][1], c[3][2], c[3][3] = x1, y1, z0
-      c[4][1], c[4][2], c[4][3] = x1, y1, z1
-    else                                                 -- west, at x0
-      c[1][1], c[1][2], c[1][3] = x0, y0, z0
-      c[2][1], c[2][2], c[2][3] = x0, y0, z1
-      c[3][1], c[3][2], c[3][3] = x0, y1, z1
-      c[4][1], c[4][2], c[4][3] = x0, y1, z0
-    end
-    local u0, u1, v0, v1 = uvRect(tile, vTop, vBot)
-    local uv = scratchUv
-    uv[1][1], uv[1][2] = u0, v1
-    uv[2][1], uv[2][2] = u1, v1
-    uv[3][1], uv[3][2] = u1, v0
-    uv[4][1], uv[4][2] = u0, v0
-
-    push(c, uv, shade)
-  end
-
-  local def = map.def
-  local tw, th = def.width * 4, def.height * 4         -- map size in tiles
-  local r = bodyOnly and 0 or RING * 4
-
-  -- true when the (ring) position lies under a connected neighbour's body
-  local function masked(px0, pz0, px1, pz1)
-    if not masks then return false end
-    for _, mk in ipairs(masks) do
-      if px1 > mk[1] and px0 < mk[3] and pz1 > mk[2] and pz0 < mk[4] then
-        return true
-      end
-    end
-    return false
-  end
-
-  -- The inclusive variant for OBJECT quads: a quad TOUCHING a neighbour
-  -- body counts as under it. The old test took the quad's center with
-  -- strict bounds, and a quad whose center sat exactly on the body's
-  -- edge line escaped the mask -- stringing stray pixel fragments of
-  -- otherwise-dropped border trees along every map seam.
-  local function maskedClosed(px0, pz0, px1, pz1)
-    if not masks then return false end
-    for _, mk in ipairs(masks) do
-      if px1 >= mk[1] and px0 <= mk[3] and pz1 >= mk[2] and pz0 <= mk[4] then
-        return true
-      end
-    end
-    return false
-  end
-
-  for ty = -r, th + r - 1 do
-    for tx = -r, tw + r - 1 do
-      -- check() (clock every call, not every 8th): the geometry
-      -- emission below is the heaviest per-cell work in the whole
-      -- build -- billboard cards, side bands, shoreline faces -- and a
-      -- sampled tick let a single cell overshoot the whole slice
-      Budget.check()
-      local k = GridKey.of(tx, ty)
-      local s, tile = S.shapeAt[k], S.tileAt[k]
-      local inBody = tx >= 0 and ty >= 0 and tx < tw and ty < th
-      if not inBody and masked(tx * 8, ty * 8, tx * 8 + 8, ty * 8 + 8) then
-        s = nil
-      end
-
-      -- Under the TREES fill the border wall is MODELLED or it is not there
-      -- (see Structures' hullRingOnly): a ring cell nothing claimed would
-      -- be a flat-topped box standing beside carved trunks, which reads as
-      -- a painted-on plateau rather than forest. Structures already stops
-      -- the ring at the carve distance; this catches the odd cell inside it
-      -- that the 2x2 grouping could not take -- a canopy whose partners
-      -- fall outside the shortened ring is left unclaimed, and one strip of
-      -- boxes along an edge is the whole artefact this avoids.
-      if not inBody and S.hideBareRing and not S.skip[k] then
-        s = nil
-      end
-
-      if s and S.skip[k] then
-        -- an object stands here; paint its synthesized ground and let the
-        -- prebuilt prism quads (appended below) carry the art
-        local g = S.ground[k]
-        if g then
-          topQuad(tx * 8, ty * 8, 0, g, 1)
-          -- the claimed tile is still ground at height 0, and water next
-          -- door still recesses below it: without the same below-ground
-          -- side bands ordinary ground emits, the two-pixel shoreline
-          -- face is a slit into the sky behind the mesh -- which is
-          -- exactly what a building plot or a sign standing at the
-          -- waterline showed. Same bands, cut from the synthesized
-          -- ground's own art
-          for _, side in ipairs(SIDES) do
-            local nh = heightAt(tx + side[1], ty + side[2])
-            if nh < 0 then
-              local d = side[3]
-              local lat = LATERAL[d]
-              local hl = lat and heightAt(tx + lat[1], ty + lat[2]) or 0
-              local hr = lat and heightAt(tx + lat[3], ty + lat[4]) or 0
-              for band = math.floor(nh / 8), -1 do
-                local y0 = math.max(nh, band * 8)
-                local y1 = math.min(0, band * 8 + 8)
-                if y1 > y0 then
-                  sideQuad(d, tx * 8, ty * 8, y0, y1, g,
-                           (band * 8 + 8) - y1, (band * 8 + 8) - y0,
-                           sideShades(hl, hr, y0, y1, y0 <= nh,
-                                      Voxel3D.FACE_SHADE[d]))
-                end
-              end
-            end
-          end
-        end
-      elseif s then
-        local run = S.runs[k]
-        local h = run and run.h or s.h
-        local x0, z0 = tx * 8, ty * 8
-
-        -- top face. A roofed volume gets a GABLE segment: the roof rises
-        -- from the facade top at the south eave to a ridge across the
-        -- footprint's middle, then falls back to the facade at the north
-        -- edge -- so the far side sits LOW. (The first cut was a shed
-        -- plane rising all the way north, which turns a building into a
-        -- ramp.) The south slope wears the structure's roof rows (ridge
-        -- art at the ridge, eaves art at the eave); the back slope
-        -- mirrors them. Exposed east/west flanks hip: their outer edge
-        -- drops toward the eave, rounding the drawn corner tiles into 45
-        -- degree corners. Flat-topped volumes wear their top rows;
-        -- everything else its own art.
-        if run and run.rise > 0 then
-          local mid = run.extent / 2
-          local function gableH(d)     -- d = rows north of the south eave
-            local t = d <= mid and d / mid or (run.extent - d) / (run.extent - mid)
-            return run.h + run.rise * math.max(0, math.min(1, t))
-          end
-          local d0 = run.front - ty                -- rows from the south edge
-          local hS = gableH(d0)
-          local hN = gableH(d0 + 1)
-          -- art by proximity to the ridge, mirrored over the back
-          local rel = 1 - math.abs(d0 + 0.5 - mid) / math.max(mid, 0.5)
-          local idx = math.min(run.roofRows - 1,
-                               math.floor((1 - rel) * run.roofRows))
-          local roofTile = map:tileAt(tx, run.north + idx)
-          local swY, seY, neY, nwY = hS, hS, hN, hN
-          if heightAt(tx - 1, ty) < run.h then     -- west flank: hip
-            swY = math.max(run.h, hS - 8)
-            nwY = math.max(run.h, hN - 8)
-          end
-          if heightAt(tx + 1, ty) < run.h then     -- east flank: hip
-            seY = math.max(run.h, hS - 8)
-            neY = math.max(run.h, hN - 8)
-          end
-          local u0, u1, v0, v1 = uvRect(roofTile, 0, 8)
-          push({ { x0, swY, z0 + 8 }, { x0 + 8, seY, z0 + 8 },
-                 { x0 + 8, neY, z0 }, { x0, nwY, z0 } },
-               { { u0, v1 }, { u1, v1 }, { u1, v0 }, { u0, v0 } }, 0.95)
-        elseif run then
-          local m = math.min(2, run.extent)
-          local topTile = map:tileAt(tx, run.north + ((ty - run.north) % m))
-          topQuad(x0, z0, h, topTile, VOLUME_TOP_SHADE)
-        else
-          local topTile = tile
-          if s.art == "upright" and s.authored then
-            -- Top art for a pinned box.  A furniture drawing is top-view
-            -- rows over floor(h/8) face-on rows the fold stands upright;
-            -- a face row's top would repeat its front art lying flat, so
-            -- it wears the nearest row above the face block instead --
-            -- the drawn tabletop (and whatever sits on it) stays on top,
-            -- and a fully-folded structure (wall, desk) tops with its
-            -- northmost row.
-            local north, front = ty, ty
-            while ty - north < 6 do
-              local bs = S.shapeAt[GridKey.of(tx, north - 1)]
-              if bs and bs.authored and bs.class == s.class then
-                north = north - 1
-              else
-                break
-              end
-            end
-            while front - ty < 6 do
-              local bs = S.shapeAt[GridKey.of(tx, front + 1)]
-              if bs and bs.authored and bs.class == s.class then
-                front = front + 1
-              else
-                break
-              end
-            end
-            local row = math.min(ty, front - math.floor(h / 8))
-            if row < north then
-              -- the whole run folded onto the face: top with the drawn
-              -- row just above it when that row is furniture too (a
-              -- bookcase wearing its shelf-top trim), else with the
-              -- run's own top row
-              local above = S.shapeAt[GridKey.of(tx, north - 1)]
-              row = (above and above.authored and above.art == "upright")
-                    and (north - 1) or north
-            end
-            topTile = S.tileAt[GridKey.of(tx, row)]
-          end
-          -- water's surface, and only water's: the recessed sheet itself,
-          -- never the ground's shoreline bands around it. A cell an object
-          -- stands on took the branch above and paints synthesized GROUND,
-          -- which is right -- a sign at the waterline stands on a plot, not
-          -- on the pond.
-          topQuad(x0, z0, h, topTile,
-                  s.art == "upright" and VOLUME_TOP_SHADE or 1,
-                  (s.class == "water") and waterPush or nil)
-        end
-
-        -- sides: 8px bands wherever the neighbour is lower. Band k spans
-        -- heights [8k, 8k+8) and shows one full tile of art; a partial
-        -- band crops the art rows to match, so nothing ever stretches.
-        for _, side in ipairs(SIDES) do
-          local nh = heightAt(tx + side[1], ty + side[2])
-          if nh < h then
-            local d = side[3]
-            -- the columns flanking this face, for the inside-corner term:
-            -- fixed for the whole face, so they are read once rather than
-            -- once per 8px band
-            local lat = LATERAL[d]
-            local hl = lat and heightAt(tx + lat[1], ty + lat[2]) or 0
-            local hr = lat and heightAt(tx + lat[3], ty + lat[4]) or 0
-            for band = math.floor(nh / 8), math.ceil(h / 8) - 1 do
-              local y0 = math.max(nh, band * 8)
-              local y1 = math.min(h, band * 8 + 8)
-              if y1 > y0 then
-                local src, shade = tile, Voxel3D.FACE_SHADE[d]
-                if run then
-                  -- fold the structure's artwork up this face: band k
-                  -- samples the map row k tiles north of the structure's
-                  -- front, clamped to its extent. The south face is the
-                  -- drawing itself (full brightness); the other sides wear
-                  -- the same rows darkened, so a building's flank matches
-                  -- its face instead of smearing one tile
-                  if d == 6 then
-                    src = map:tileAt(tx, math.min(run.front,
-                                                  run.north + band))
-                  else
-                    src = map:tileAt(tx, math.max(run.north,
-                                                  run.front - band))
-                  end
-                  if d == 5 then shade = 1 end
-                elseif s.art == "upright" then
-                  -- profile-authored upright (a pinned wall or furniture
-                  -- box): fold the drawing up the face, band 0 the
-                  -- structure's southmost same-class row and higher bands
-                  -- the rows north of it, repeating past the top.  The
-                  -- south face is the drawing itself (full brightness);
-                  -- flanks and back wear the same front stack darkened, so
-                  -- a desk's side matches its face instead of smearing a
-                  -- different jumble per row.
-                  if d == 5 then shade = 1 end
-                  local front = ty
-                  while front < ty + 6 do
-                    local fs2 = S.shapeAt[GridKey.of(tx, front + 1)]
-                    if fs2 and fs2.authored and fs2.class == s.class then
-                      front = front + 1
-                    else
-                      break
-                    end
-                  end
-                  local fk = GridKey.of(tx, front - band)
-                  local fs = S.shapeAt[fk]
-                  if fs and fs.authored and fs.class == s.class then
-                    src = S.tileAt[fk]
-                  end
-                end
-                sideQuad(d, x0, z0, y0, y1, src,
-                         (band * 8 + 8) - y1, (band * 8 + 8) - y0,
-                         sideShades(hl, hr, y0, y1, y0 <= nh, shade))
-              end
-            end
-          end
-        end
-      end
-    end
-  end
-
-  -- Prebuilt quads from Structures (per-pixel voxel props, lathed
-  -- columns) plus the round-tree stamps expanded in place. Keep rules,
-  -- by the quad's own extent:
-  --   body-only   the quad must overlap the OPEN body interval -- a
-  --               neighbour's ring props must not march past its edge
-  --               into this map, and a quad lying exactly ON the edge
-  --               plane would z-fight the map that owns that plane.
-  --   full        anything overlapping the body stays whole (props that
-  --               straddle the edge no longer shed their outer half);
-  --               pure ring quads drop when they touch a neighbour body
-  --               (maskedClosed), which is what strings of seam pixels
-  --               were: fragments of dropped border trees whose centers
-  --               sat exactly on the boundary line.
-  local bw, bh = tw * 8, th * 8
-  local function keepQuad(x0, z0, x1, z1)
-    local overBody = x1 > 0 and x0 < bw and z1 > 0 and z0 < bh
-    if bodyOnly then return overBody end
-    return overBody or not maskedClosed(x0, z0, x1, z1)
-  end
-
-  -- A face lying EXACTLY on a body boundary plane is ambiguous to the
-  -- rect tests above: a body structure's outward facade (a Saffron row
-  -- house whose front row is the map's last row, its south wall on the
-  -- shared plane with Route 6) and the inward face of a ring scrap
-  -- occupy the same degenerate rect, and the strict overBody plus the
-  -- closed mask dropped BOTH -- which is why those facades were missing.
-  -- The winding tells them apart: a face pointing AWAY from the body
-  -- belongs to this map's own edge-row structure and nothing in the
-  -- neighbour will ever draw that plane, so it stays; a face pointing
-  -- INTO the body is the scrap the mask rules exist to kill, and falls
-  -- through to them.
-  local function outwardOnEdge(q, x0, z0, x1, z1)
-    if z0 == z1 and (z0 == 0 or z0 == bh) and x1 > 0 and x0 < bw then
-      local nz = (q[2][1] - q[1][1]) * (q[3][2] - q[1][2])
-                 - (q[2][2] - q[1][2]) * (q[3][1] - q[1][1])
-      return (z0 == bh and nz > 0) or (z0 == 0 and nz < 0)
-    end
-    if x0 == x1 and (x0 == 0 or x0 == bw) and z1 > 0 and z0 < bh then
-      local nx = (q[2][2] - q[1][2]) * (q[3][3] - q[1][3])
-                 - (q[2][3] - q[1][3]) * (q[3][2] - q[1][2])
-      return (x0 == bw and nx > 0) or (x0 == 0 and nx < 0)
-    end
-    return false
-  end
-
-  local scUV = { { 0, 0 }, { 0, 0 }, { 0, 0 }, { 0, 0 } }
-  local scObject = { { 0, 0, 0 }, { 0, 0, 0 },
-                     { 0, 0, 0 }, { 0, 0, 0 } }
-  local function quadUV(q)
-    if q.uv then return q.uv end
-    for i = 1, 4 do
-      scUV[i][1], scUV[i][2] = q.u, q.v
-    end
-    return scUV
-  end
-
-  for _, q in ipairs(S.objectQuads) do
-    Budget.tick()
-    -- Building stamps retain their immutable template and placement offset;
-    -- materialize into reusable corners before bounds, culling, and push.
-    local source, drawQ = q, q
-    if q.localQ then
-      source = q.localQ
-      local ox, oz = q.offsetX, q.offsetZ
-      for i = 1, 4 do
-        local c, d = source[i], scObject[i]
-        d[1], d[2], d[3] = c[1] + ox, c[2], c[3] + oz
-      end
-      drawQ = scObject
-    end
-    local x0 = math.min(drawQ[1][1], drawQ[2][1], drawQ[3][1], drawQ[4][1])
-    local x1 = math.max(drawQ[1][1], drawQ[2][1], drawQ[3][1], drawQ[4][1])
-    local z0 = math.min(drawQ[1][3], drawQ[2][3], drawQ[3][3], drawQ[4][3])
-    local z1 = math.max(drawQ[1][3], drawQ[2][3], drawQ[3][3], drawQ[4][3])
-    -- q.own: a body-anchored structure's own quad (a building placed by
-    -- Buildings.build, whose scan never leaves the body). Exempt from
-    -- the edge keep-rules entirely: its eave legitimately overhangs the
-    -- boundary plane into the neighbour's airspace, and no variant of
-    -- the neighbour will ever draw that geometry
-    if q.own or outwardOnEdge(drawQ, x0, z0, x1, z1)
-       or keepQuad(x0, z0, x1, z1) then
-      push({ drawQ[1], drawQ[2], drawQ[3], drawQ[4] },
-           quadUV(q), groundShades(q, q.shade))
-    end
-  end
-
-  -- round-tree stamps: the shared hull template translated per cell,
-  -- through reusable scratch corners so expansion allocates nothing.
-  -- A hull spans at most its own footprint -- one 16px cell unless the
-  -- stamp carries a wider radius (the 2x2-cell canopy groups) -- so one
-  -- rect test usually answers for the whole stamp: strictly interior
-  -- stamps keep every quad, ring stamps buried under a neighbour body
-  -- (or, body-only, ring stamps full stop) skip without touching their
-  -- quads.
-  --
-  -- A stamp is ONE tree, and the tree is atomic: the mask must never
-  -- cull its quads piecemeal. A stamp straddling the transition line
-  -- (partly under a neighbour body, partly on this map's ring) used to
-  -- fall through to per-quad keepQuad, which dropped every quad touching
-  -- the mask -- leaving a tree cut in half along the seam. The trunk
-  -- (the stamp centre) decides instead: a trunk under a neighbour body
-  -- is a tree that would rise through the neighbour's flat ground, so
-  -- the whole stamp goes; a trunk on this map's side keeps the whole
-  -- stamp, canopy overhang and all -- the overhang is above the
-  -- neighbour's ground, which is what a tree at a road edge does, and
-  -- the depth buffer sorts it against the neighbour's own geometry.
-  local sc = { { 0, 0, 0 }, { 0, 0, 0 }, { 0, 0, 0 }, { 0, 0, 0 } }
-  for _, st in ipairs(S.roundStamps or {}) do
-    local mx, mz = st.mx, st.mz
-    local sr = st.r or 8
-    local sx0, sz0, sx1, sz1 = mx - sr, mz - sr, mx + sr, mz + sr
-    local skipAll
-    if bodyOnly then
-      -- A neighbour contributes only its body. Same atomic rule as the
-      -- full variant, against the body rect instead of the mask rects:
-      -- the trunk (stamp centre) decides, so a body tree whose canopy
-      -- overhangs the body edge keeps every quad (a tree at the forest
-      -- edge legitimately overhangs the seam), and a ring tree whose
-      -- canopy pokes INTO the body is dropped whole -- its half-drawn
-      -- canopy used to float over the neighbour's ground with no trunk
-      -- (the per-quad slice), which read as a tree cut in half.
-      local centreInBody = mx >= 0 and mx <= bw and mz >= 0 and mz <= bh
-      skipAll = not centreInBody
-    else
-      -- Full variant: the mask rects are where connected neighbour
-      -- BODIES sit. A trunk under one is a tree that would rise through
-      -- the neighbour's flat ground -- the whole stamp goes. A trunk on
-      -- this map's side keeps the whole stamp, canopy overhang and all:
-      -- the overhang is above the neighbour's ground, which is what a
-      -- tree at a road edge does, and the depth buffer sorts it against
-      -- the neighbour's own geometry.
-      local centreInMask = false
-      if masks then
-        for _, mk in ipairs(masks) do
-          if mx >= mk[1] and mx <= mk[3] and mz >= mk[2] and mz <= mk[4] then
-            centreInMask = true
-            break
-          end
-        end
-      end
-      skipAll = centreInMask
-    end
-    if not skipAll then
-      for _, q in ipairs(st.quads) do
-        Budget.tick()
-        for i = 1, 4 do
-          local c, s2 = q[i], sc[i]
-          s2[1] = c[1] + mx
-          s2[2] = c[2]
-          s2[3] = c[3] + mz
-        end
-        push(sc, quadUV(q), groundShades(sc, q.shade))
-      end
-    end
-  end
-end
+-- GeometryBuilder emits the same sink protocol for every build path.
 
 -- The raw geometry for `map`: (vertex list, triangle index list, quad
 -- count). Synchronous and GPU-free -- the headless suite and the probes
@@ -834,7 +191,7 @@ end
 function ChunkMesher.geometry(map, bodyOnly, masks, split)
   local sink = newTableSink()
   local waterSink = split and newTableSink() or nil
-  runGeometry(map, bodyOnly, masks, sink, waterSink)
+  GeometryBuilder.emit(map, bodyOnly, masks, sink, waterSink)
   if not waterSink then return sink.results() end
   local v, i, n = sink.results()
   local wv, wi, wn = waterSink.results()
@@ -851,7 +208,7 @@ end
 function ChunkMesher.build(map, bodyOnly, masks, split)
   local sink = newSink()
   local waterSink = split and newSink() or nil
-  runGeometry(map, bodyOnly, masks, sink, waterSink)
+  GeometryBuilder.emit(map, bodyOnly, masks, sink, waterSink)
   return sink.finish(), waterSink and waterSink.finish() or nil
 end
 
@@ -950,15 +307,7 @@ local releaseEntry = Runtime.releaseEntry
 
 -- ---------------------------------------------------------- async builds
 
-local jobs = {}       -- FIFO of pending jobs
-local jobIndex = {}   -- "id:slot" -> job
-local completion = {} -- "id:slot" -> complete|failed|cancelled
-
 local clock = (love and love.timer and love.timer.getTime) or os.clock
-
-local function jobKey(id, slot)
-  return id .. ":" .. slot
-end
 
 local DebugOverlay = nil
 local function debugNote(fmt, ...)
@@ -969,7 +318,7 @@ local function debugNote(fmt, ...)
   if DebugOverlay then DebugOverlay.note(fmt, ...) end
 end
 local function finishJob(job, ok, err)
-  local key = jobKey(job.id, job.slot)
+  local key = Queue.key(job.id, job.slot)
   local jobMs = nil
   if job.queuedAt and love and love.timer and love.timer.getTime then
     jobMs = math.floor((love.timer.getTime() - job.queuedAt) * 1000 + 0.5)
@@ -993,14 +342,7 @@ local function finishJob(job, ok, err)
                         job.maxGapMs or 0, job.overshoots or 0)
     end
   end
-  jobIndex[key] = nil
-  completion[key] = ok and "complete" or "failed"
-  for i, j in ipairs(jobs) do
-    if j == job then
-      table.remove(jobs, i)
-      break
-    end
-  end
+  Queue.finish(job, ok)
   if not ok then
     -- name the reason: in a real session a lost build is a black map
     print("[warn] voxel mesh build failed for " .. tostring(job.id)
@@ -1181,7 +523,7 @@ local function runJob(job)
   job.phase = "geometry"
   local sink = newSink()
   local waterSink = newSink()
-  runGeometry(map, job.slot == "body", job.masks, sink, waterSink)
+  GeometryBuilder.emit(map, job.slot == "body", job.masks, sink, waterSink)
   local mesh = sink.finish()
   local water = waterSink.finish()
   if MeshCache.available() then
@@ -1216,7 +558,7 @@ end
 function ChunkMesher.buildGeometryData(map, bodyOnly, masks)
   local sink = newSink()
   local waterSink = newSink()
-  runGeometry(map, bodyOnly, masks, sink, waterSink)
+  GeometryBuilder.emit(map, bodyOnly, masks, sink, waterSink)
   local okFlat, flat = pcall(flattenAux, map)
   local tb, tn, ti, tm = sink.buffer()
   local wb, wn, wi, wm = waterSink.buffer()
@@ -1301,20 +643,18 @@ function ChunkMesher.request(map, bodyOnly, masks, urgent, force)
   end
   local stale = c.stale and (c.stale[slot] or c.stale.aux)
   if c[slot] ~= nil and not force and not stale then return c[slot] or nil end
-  local key = jobKey(map.id, slot)
-  local job = jobIndex[key]
-  if force then completion[key] = nil end
+  local job = Queue.find(map.id, slot)
   if not job then
     job = { id = map.id, map = map, slot = slot, masks = masks,
             urgent = urgent or false, prebuild = force or false,
             gen = gen[map.id] or 0,
             queuedAt = love and love.timer and love.timer.getTime
                        and love.timer.getTime() or nil }
-    jobIndex[key] = job
-    jobs[#jobs + 1] = job
+    Queue.enqueue(job, force)
   else
     if urgent then job.urgent = true end
     if force then job.prebuild = true end
+    if force then Queue.enqueue(job, true) end
   end
   return (c and c[slot]) or nil
 end
@@ -1333,38 +673,34 @@ function ChunkMesher.requestMapId(mapId, bodyOnly, masks, urgent, force, loader)
     c.stale.aux = true
     c.stale[slot] = true
   end
-  local key = jobKey(mapId, slot)
-  local job = jobIndex[key]
-  if force then completion[key] = nil end
+  local job = Queue.find(mapId, slot)
   if not job then
     job = { id = mapId, loader = loader, slot = slot, masks = masks,
             urgent = urgent or false, prebuild = force or false,
             gen = gen[mapId] or 0,
             queuedAt = love and love.timer and love.timer.getTime
                        and love.timer.getTime() or nil }
-    jobIndex[key] = job
-    jobs[#jobs + 1] = job
+    Queue.enqueue(job, force)
   else
     if urgent then job.urgent = true end
     if force then job.prebuild = true end
+    if force then Queue.enqueue(job, true) end
   end
   return (c and c[slot]) or nil
 end
 
 function ChunkMesher.pending()
-  return #jobs
+  return Queue.pending()
 end
 
 -- A precise queue probe used by cooperative tooling. It does not inspect or
 -- retain the mesh; it only answers whether a slot is still in flight.
 function ChunkMesher.jobPending(mapId, bodyOnly)
-  return jobIndex[jobKey(mapId, bodyOnly and "body" or "full")] ~= nil
+  return Queue.jobPending(mapId, bodyOnly and "body" or "full")
 end
 
 function ChunkMesher.jobStatus(mapId, bodyOnly)
-  local key = jobKey(mapId, bodyOnly and "body" or "full")
-  if jobIndex[key] then return "pending" end
-  return completion[key]
+  return Queue.status(mapId, bodyOnly and "body" or "full")
 end
 
 -- Release a completed prebuild map immediately. Unlike invalidate(), this
@@ -1377,15 +713,7 @@ function ChunkMesher.release(mapId)
   -- Cancellation is generation-based, but also remove queued jobs so a
   -- cancelled prebuild cannot keep a map/Structures graph alive in a closure.
   gen[mapId] = (gen[mapId] or 0) + 1
-  for i = #jobs, 1, -1 do
-    local job = jobs[i]
-    if job.id == mapId then
-      local key = jobKey(job.id, job.slot)
-      jobIndex[key] = nil
-      completion[key] = "cancelled"
-      table.remove(jobs, i)
-    end
-  end
+  Queue.removeIf(function(job) return job.id == mapId end, "cancelled")
   Structures.invalidate(mapId)
   return had
 end
@@ -1409,14 +737,8 @@ ChunkMesher.IDLE_SLICE = 0.005
 ChunkMesher.COVERED_SLICE = 0.050
 
 function ChunkMesher.pump(covered)
-  if #jobs == 0 then return end
-  local pick = jobs[1]
-  for _, j in ipairs(jobs) do
-    if j.urgent then
-      pick = j
-      break
-    end
-  end
+  if Queue.pending() == 0 then return end
+  local pick = Queue.pick(true)
   local slice = covered and ChunkMesher.COVERED_SLICE
                 or (pick.urgent and ChunkMesher.URGENT_SLICE
                     or ChunkMesher.IDLE_SLICE)
@@ -1458,14 +780,8 @@ function ChunkMesher.pump(covered)
     else
       return   -- slice spent mid-build; resume next frame
     end
-    if clock() >= deadline or #jobs == 0 then return end
-    pick = jobs[1]
-    for _, j in ipairs(jobs) do
-      if j.urgent then
-        pick = j
-        break
-      end
-    end
+    if clock() >= deadline or Queue.pending() == 0 then return end
+    pick = Queue.pick(true)
   end
 end
 
@@ -1541,8 +857,7 @@ function ChunkMesher.get(map, bodyOnly, masks)
         c.stale = nil
       end
     end
-    local key = jobKey(map.id, slot)
-    local job = jobIndex[key]
+    local job = Queue.find(map.id, slot)
     if job then finishJob(job, true) end
   end
   return c[slot] or nil
@@ -1616,13 +931,7 @@ function ChunkMesher.refresh(mapId)
   MeshCache.invalidate(mapId)
   Structures.invalidate(mapId)
   gen[mapId] = (gen[mapId] or 0) + 1
-  for i = #jobs, 1, -1 do
-    local job = jobs[i]
-    if job.id == mapId then
-      jobIndex[jobKey(job.id, job.slot)] = nil
-      table.remove(jobs, i)
-    end
-  end
+  Queue.removeIf(function(job) return job.id == mapId end)
   -- false-cached slots count as stale too: a retry after a failed build
   -- is exactly a rebuild
   c.stale = { aux = true,
@@ -1646,12 +955,10 @@ local prevLive = {}
 function ChunkMesher.setLive(live)
   prevLive = Runtime.evict({
     cache = cache,
-    jobs = jobs,
+    queue = Queue,
     live = live,
     previous = prevLive,
     generations = gen,
-    index = jobIndex,
-    completion = completion,
     onEvict = function(id) Structures.invalidate(id) end,
   })
 end
@@ -1673,15 +980,9 @@ function ChunkMesher.invalidate(mapId)
     cache = {}
     for id in pairs(gen) do gen[id] = gen[id] + 1 end
   end
-  for i = #jobs, 1, -1 do
-    local job = jobs[i]
-    if mapId == nil or job.id == mapId then
-      local key = jobKey(job.id, job.slot)
-      jobIndex[key] = nil
-      completion[key] = "cancelled"
-      table.remove(jobs, i)
-    end
-  end
+  Queue.removeIf(function(job)
+    return mapId == nil or job.id == mapId
+  end, "cancelled")
 end
 
 -- The engine fires every registered invalidator at boot too: the mod
