+++ lib/GeometryBuilder.lua
@@ -1,0 +1,636 @@
+-- Pure map-to-quad stream construction for ChunkMesher.
+--
+-- This module knows the authored Structures result and voxel math, but not
+-- GPU objects, cache storage, or queue state. Sinks are injected by the
+-- caller, so the same builder serves headless geometry, serial mesh builds,
+-- and threaded-prebuild stream generation.
+local V = ...
+
+local Structures = V.require("Structures")
+local Voxel3D = V.require("Voxel3D")
+local Budget = V.require("BuildBudget")
+local GridKey = V.require("GridKey")
+
+local GeometryBuilder = {}
+
+local RING = 3
+local INSET = 0.02
+local VOLUME_TOP_SHADE = 0.85
+local SIDES = {
+  { 1, 0, 1 },
+  { -1, 0, 2 },
+  { 0, 1, 5 },
+  { 0, -1, 6 },
+}
+
+function GeometryBuilder.emit(map, bodyOnly, masks, sink, waterSink)
+  local push = sink.push
+  local waterPush = waterSink and waterSink.push or nil
+  local tileset = map.tileset
+  local S = Structures.forMap(map)
+  local perRow = tileset.tilesPerRow or 16
+  local atlasW = tileset.imageWidth or (perRow * 8)
+  local atlasH = tileset.imageHeight or 48
+
+  local function heightAt(tx, ty)
+    local k = GridKey.of(tx, ty)
+    if S.skip[k] then return 0 end
+    local run = S.runs[k]
+    if run then return run.h end
+    local s = S.shapeAt[k]
+    return s and s.h or 0
+  end
+
+  -- one atlas-rect UV, optionally cropped to art rows [vTop, vBot] of 8
+  local function uvRect(tile, vTop, vBot)
+    local ax = (tile % perRow) * 8
+    local ay = math.floor(tile / perRow) * 8
+    local vi = math.min(INSET, (vBot - vTop) / 4)
+    return (ax + INSET) / atlasW, (ax + 8 - INSET) / atlasW,
+           (ay + vTop + vi) / atlasH, (ay + vBot - vi) / atlasH
+  end
+
+  -- ------------------------------------------------------ ambient occlusion
+  --
+  -- Ambient light is what reaches a surface from the sky at large, so it is
+  -- blocked by how much geometry crowds a point rather than by where the
+  -- sun happens to be -- which makes it the exact complement of the shadow
+  -- pass, and the reason both are worth having. The shadow map draws the
+  -- long directional shadow a building throws; this draws the dark seam in
+  -- every corner the sky cannot see into, at every scale finer than a
+  -- shadow map texel.
+  --
+  -- Baked per vertex, the classic voxel way: each corner counts the
+  -- neighbours that crowd it and steps down once per neighbour, and the
+  -- rasteriser interpolates the steps into a smooth falloff. Costs exactly
+  -- nothing at draw time, and it is resolution-independent -- a screen
+  -- space pass would blur across the pixel grid this whole mode is built
+  -- to keep crisp.
+  --
+  -- (What was here before was a one-directional contact shadow keyed to a
+  -- sun in the northwest: two neighbours, one corner, top faces only.)
+
+  -- Intensity. Both terms below are DARKENING amounts rather than
+  -- multipliers, so this one number scales the whole effect: 1.0 is the
+  -- barely-there first cut, and everything is expressed against it.
+  local AO_STRENGTH = 2.4
+  local AO_STEP = 0.09 * AO_STRENGTH      -- per crowding neighbour, max 3
+  local AO_EDGE = 1 - 0.14 * AO_STRENGTH  -- creases / corners on a face
+  local AO_GROUND = 0.12 * AO_STRENGTH    -- a prop's contact with the floor
+  local AO_RISE = 6                       -- px over which the floor lets go
+  local AO_FLOOR = 0.25                   -- never let a vertex reach black
+
+  -- Both sinks copy a per-corner shade straight out into the vertex stream
+  -- and keep no reference, so these two scratch rows are reused for every
+  -- quad on the map rather than allocating a table per face -- a route
+  -- builds a few hundred thousand of them.
+  local aoTop = { 0, 0, 0, 0 }
+  local aoSide = { 0, 0, 0, 0 }
+
+  -- A top face's four corners, each occluded by the three cells that touch
+  -- it: two edge neighbours and the diagonal between them.
+  local function aoShades(tx, ty, h, shade)
+    local n = heightAt(tx, ty - 1) > h
+    local s = heightAt(tx, ty + 1) > h
+    local e = heightAt(tx + 1, ty) > h
+    local w = heightAt(tx - 1, ty) > h
+    local nw = heightAt(tx - 1, ty - 1) > h
+    local ne = heightAt(tx + 1, ty - 1) > h
+    local sw = heightAt(tx - 1, ty + 1) > h
+    local se = heightAt(tx + 1, ty + 1) > h
+    if not (n or s or e or w or nw or ne or sw or se) then return shade end
+    local function corner(a, b, d)
+      local k = 0
+      if a then k = k + 1 end
+      if b then k = k + 1 end
+      -- a diagonal wedged behind both of its edges adds nothing: the
+      -- corner is already as enclosed as it can get, and counting it
+      -- again is what turns an ordinary inside corner black
+      if d and not (a and b) then k = k + 1 end
+      -- floored, so cranking AO_STRENGTH deepens the seams instead of
+      -- punching holes of pure black through the world
+      return shade * math.max(AO_FLOOR, 1 - AO_STEP * k)
+    end
+    -- corners in topQuad order: NW, NE, SE, SW
+    aoTop[1], aoTop[2] = corner(n, w, nw), corner(n, e, ne)
+    aoTop[3], aoTop[4] = corner(s, e, se), corner(s, w, sw)
+    return aoTop
+  end
+
+  -- The same idea on an upright face, where the crowding is of two kinds:
+  -- the CREASE it rises out of (the band sitting on the ground, or on
+  -- whatever lower neighbour exposed the face) and the INSIDE CORNERS
+  -- where the columns flanking it stand proud of the band. `hl`/`hr` are
+  -- those flanking heights in FACE order -- left then right as seen from
+  -- outside, per LATERAL below -- so the shades line up with sideQuad's
+  -- corners without the caller thinking about compass directions.
+  local LATERAL = {
+    [1] = { 0, 1, 0, -1 },    -- east face:  left south, right north
+    [2] = { 0, -1, 0, 1 },    -- west face:  left north, right south
+    [5] = { -1, 0, 1, 0 },    -- south face: left west,  right east
+    [6] = { 1, 0, -1, 0 },    -- north face: left east,  right west
+  }
+  -- Ground contact for the prebuilt prop quads -- the per-pixel plants,
+  -- signs and lone trees, and the round-tree stamps. Those arrive from
+  -- Structures already finished, so the neighbour counting above has no
+  -- columns to count. What it CAN say is that the ground plane itself
+  -- blocks half the sky, so the closer a voxel sits to it the less ambient
+  -- light reaches it -- which is what plants a prop on the floor instead
+  -- of leaving it looking pasted over the top.
+  local aoProp = { 0, 0, 0, 0 }
+  local function groundShades(c, shade)
+    if type(shade) == "table" then return shade end
+    local y1, y2, y3, y4 = c[1][2], c[2][2], c[3][2], c[4][2]
+    if math.min(y1, y2, y3, y4) >= AO_RISE then return shade end
+    for i = 1, 4 do
+      local t = c[i][2] / AO_RISE
+      aoProp[i] = shade * (t >= 1 and 1 or (1 - AO_GROUND * (1 - t)))
+    end
+    return aoProp
+  end
+
+  local AO_CORNER = math.max(AO_FLOOR, AO_EDGE * AO_EDGE)  -- crease AND flank
+  local function sideShades(hl, hr, y0, y1, crease, shade)
+    if not (crease or hl > y0 or hr > y0) then return shade end
+    -- corners run bottom-left, bottom-right, top-right, top-left
+    local base = crease and AO_EDGE or 1
+    aoSide[1] = shade * (hl > y0 and (crease and AO_CORNER or AO_EDGE) or base)
+    aoSide[2] = shade * (hr > y0 and (crease and AO_CORNER or AO_EDGE) or base)
+    aoSide[3] = shade * (hr > y1 and AO_EDGE or 1)
+    aoSide[4] = shade * (hl > y1 and AO_EDGE or 1)
+    return aoSide
+  end
+
+  local scratchC = { {0,0,0}, {0,0,0}, {0,0,0}, {0,0,0} }
+  local scratchUv = { {0,0}, {0,0}, {0,0}, {0,0} }
+
+  -- `to` routes the quad somewhere other than the main sink -- the water
+  -- surface is the only caller that ever does (see runGeometry's header).
+  local function topQuad(x0, z0, h, tile, shade, to)
+    local u0, u1, v0, v1 = uvRect(tile, 0, 8)
+    local c = scratchC
+    c[1][1], c[1][2], c[1][3] = x0, h, z0
+    c[2][1], c[2][2], c[2][3] = x0 + 8, h, z0
+    c[3][1], c[3][2], c[3][3] = x0 + 8, h, z0 + 8
+    c[4][1], c[4][2], c[4][3] = x0, h, z0 + 8
+
+    local uv = scratchUv
+    uv[1][1], uv[1][2] = u0, v0
+    uv[2][1], uv[2][2] = u1, v0
+    uv[3][1], uv[3][2] = u1, v1
+    uv[4][1], uv[4][2] = u0, v1
+
+    ;(to or push)(c, uv, aoShades(x0 / 8, z0 / 8, h, shade))
+  end
+
+  -- vertical quad for face direction `d` of the tile column at (x0, z0),
+  -- spanning heights [y0, y1] and showing art rows [vTop, vBot] of `tile`.
+  -- Corners run bottom-left, bottom-right, top-right, top-left as seen
+  -- from outside; u follows +X on the north/south faces so a door or sign
+  -- never draws mirrored.
+  local function sideQuad(d, x0, z0, y0, y1, tile, vTop, vBot, shade)
+    local x1, z1 = x0 + 8, z0 + 8
+    local c = scratchC
+    if d == 5 then                                       -- south, at z1
+      c[1][1], c[1][2], c[1][3] = x0, y0, z1
+      c[2][1], c[2][2], c[2][3] = x1, y0, z1
+      c[3][1], c[3][2], c[3][3] = x1, y1, z1
+      c[4][1], c[4][2], c[4][3] = x0, y1, z1
+    elseif d == 6 then                                   -- north, at z0
+      c[1][1], c[1][2], c[1][3] = x1, y0, z0
+      c[2][1], c[2][2], c[2][3] = x0, y0, z0
+      c[3][1], c[3][2], c[3][3] = x0, y1, z0
+      c[4][1], c[4][2], c[4][3] = x1, y1, z0
+    elseif d == 1 then                                   -- east, at x1
+      c[1][1], c[1][2], c[1][3] = x1, y0, z1
+      c[2][1], c[2][2], c[2][3] = x1, y0, z0
+      c[3][1], c[3][2], c[3][3] = x1, y1, z0
+      c[4][1], c[4][2], c[4][3] = x1, y1, z1
+    else                                                 -- west, at x0
+      c[1][1], c[1][2], c[1][3] = x0, y0, z0
+      c[2][1], c[2][2], c[2][3] = x0, y0, z1
+      c[3][1], c[3][2], c[3][3] = x0, y1, z1
+      c[4][1], c[4][2], c[4][3] = x0, y1, z0
+    end
+    local u0, u1, v0, v1 = uvRect(tile, vTop, vBot)
+    local uv = scratchUv
+    uv[1][1], uv[1][2] = u0, v1
+    uv[2][1], uv[2][2] = u1, v1
+    uv[3][1], uv[3][2] = u1, v0
+    uv[4][1], uv[4][2] = u0, v0
+
+    push(c, uv, shade)
+  end
+
+  local def = map.def
+  local tw, th = def.width * 4, def.height * 4         -- map size in tiles
+  local r = bodyOnly and 0 or RING * 4
+
+  -- true when the (ring) position lies under a connected neighbour's body
+  local function masked(px0, pz0, px1, pz1)
+    if not masks then return false end
+    for _, mk in ipairs(masks) do
+      if px1 > mk[1] and px0 < mk[3] and pz1 > mk[2] and pz0 < mk[4] then
+        return true
+      end
+    end
+    return false
+  end
+
+  -- The inclusive variant for OBJECT quads: a quad TOUCHING a neighbour
+  -- body counts as under it. The old test took the quad's center with
+  -- strict bounds, and a quad whose center sat exactly on the body's
+  -- edge line escaped the mask -- stringing stray pixel fragments of
+  -- otherwise-dropped border trees along every map seam.
+  local function maskedClosed(px0, pz0, px1, pz1)
+    if not masks then return false end
+    for _, mk in ipairs(masks) do
+      if px1 >= mk[1] and px0 <= mk[3] and pz1 >= mk[2] and pz0 <= mk[4] then
+        return true
+      end
+    end
+    return false
+  end
+
+  for ty = -r, th + r - 1 do
+    for tx = -r, tw + r - 1 do
+      -- check() (clock every call, not every 8th): the geometry
+      -- emission below is the heaviest per-cell work in the whole
+      -- build -- billboard cards, side bands, shoreline faces -- and a
+      -- sampled tick let a single cell overshoot the whole slice
+      Budget.check()
+      local k = GridKey.of(tx, ty)
+      local s, tile = S.shapeAt[k], S.tileAt[k]
+      local inBody = tx >= 0 and ty >= 0 and tx < tw and ty < th
+      if not inBody and masked(tx * 8, ty * 8, tx * 8 + 8, ty * 8 + 8) then
+        s = nil
+      end
+
+      -- Under the TREES fill the border wall is MODELLED or it is not there
+      -- (see Structures' hullRingOnly): a ring cell nothing claimed would
+      -- be a flat-topped box standing beside carved trunks, which reads as
+      -- a painted-on plateau rather than forest. Structures already stops
+      -- the ring at the carve distance; this catches the odd cell inside it
+      -- that the 2x2 grouping could not take -- a canopy whose partners
+      -- fall outside the shortened ring is left unclaimed, and one strip of
+      -- boxes along an edge is the whole artefact this avoids.
+      if not inBody and S.hideBareRing and not S.skip[k] then
+        s = nil
+      end
+
+      if s and S.skip[k] then
+        -- an object stands here; paint its synthesized ground and let the
+        -- prebuilt prism quads (appended below) carry the art
+        local g = S.ground[k]
+        if g then
+          topQuad(tx * 8, ty * 8, 0, g, 1)
+          -- the claimed tile is still ground at height 0, and water next
+          -- door still recesses below it: without the same below-ground
+          -- side bands ordinary ground emits, the two-pixel shoreline
+          -- face is a slit into the sky behind the mesh -- which is
+          -- exactly what a building plot or a sign standing at the
+          -- waterline showed. Same bands, cut from the synthesized
+          -- ground's own art
+          for _, side in ipairs(SIDES) do
+            local nh = heightAt(tx + side[1], ty + side[2])
+            if nh < 0 then
+              local d = side[3]
+              local lat = LATERAL[d]
+              local hl = lat and heightAt(tx + lat[1], ty + lat[2]) or 0
+              local hr = lat and heightAt(tx + lat[3], ty + lat[4]) or 0
+              for band = math.floor(nh / 8), -1 do
+                local y0 = math.max(nh, band * 8)
+                local y1 = math.min(0, band * 8 + 8)
+                if y1 > y0 then
+                  sideQuad(d, tx * 8, ty * 8, y0, y1, g,
+                           (band * 8 + 8) - y1, (band * 8 + 8) - y0,
+                           sideShades(hl, hr, y0, y1, y0 <= nh,
+                                      Voxel3D.FACE_SHADE[d]))
+                end
+              end
+            end
+          end
+        end
+      elseif s then
+        local run = S.runs[k]
+        local h = run and run.h or s.h
+        local x0, z0 = tx * 8, ty * 8
+
+        -- top face. A roofed volume gets a GABLE segment: the roof rises
+        -- from the facade top at the south eave to a ridge across the
+        -- footprint's middle, then falls back to the facade at the north
+        -- edge -- so the far side sits LOW. (The first cut was a shed
+        -- plane rising all the way north, which turns a building into a
+        -- ramp.) The south slope wears the structure's roof rows (ridge
+        -- art at the ridge, eaves art at the eave); the back slope
+        -- mirrors them. Exposed east/west flanks hip: their outer edge
+        -- drops toward the eave, rounding the drawn corner tiles into 45
+        -- degree corners. Flat-topped volumes wear their top rows;
+        -- everything else its own art.
+        if run and run.rise > 0 then
+          local mid = run.extent / 2
+          local function gableH(d)     -- d = rows north of the south eave
+            local t = d <= mid and d / mid or (run.extent - d) / (run.extent - mid)
+            return run.h + run.rise * math.max(0, math.min(1, t))
+          end
+          local d0 = run.front - ty                -- rows from the south edge
+          local hS = gableH(d0)
+          local hN = gableH(d0 + 1)
+          -- art by proximity to the ridge, mirrored over the back
+          local rel = 1 - math.abs(d0 + 0.5 - mid) / math.max(mid, 0.5)
+          local idx = math.min(run.roofRows - 1,
+                               math.floor((1 - rel) * run.roofRows))
+          local roofTile = map:tileAt(tx, run.north + idx)
+          local swY, seY, neY, nwY = hS, hS, hN, hN
+          if heightAt(tx - 1, ty) < run.h then     -- west flank: hip
+            swY = math.max(run.h, hS - 8)
+            nwY = math.max(run.h, hN - 8)
+          end
+          if heightAt(tx + 1, ty) < run.h then     -- east flank: hip
+            seY = math.max(run.h, hS - 8)
+            neY = math.max(run.h, hN - 8)
+          end
+          local u0, u1, v0, v1 = uvRect(roofTile, 0, 8)
+          push({ { x0, swY, z0 + 8 }, { x0 + 8, seY, z0 + 8 },
+                 { x0 + 8, neY, z0 }, { x0, nwY, z0 } },
+               { { u0, v1 }, { u1, v1 }, { u1, v0 }, { u0, v0 } }, 0.95)
+        elseif run then
+          local m = math.min(2, run.extent)
+          local topTile = map:tileAt(tx, run.north + ((ty - run.north) % m))
+          topQuad(x0, z0, h, topTile, VOLUME_TOP_SHADE)
+        else
+          local topTile = tile
+          if s.art == "upright" and s.authored then
+            -- Top art for a pinned box.  A furniture drawing is top-view
+            -- rows over floor(h/8) face-on rows the fold stands upright;
+            -- a face row's top would repeat its front art lying flat, so
+            -- it wears the nearest row above the face block instead --
+            -- the drawn tabletop (and whatever sits on it) stays on top,
+            -- and a fully-folded structure (wall, desk) tops with its
+            -- northmost row.
+            local north, front = ty, ty
+            while ty - north < 6 do
+              local bs = S.shapeAt[GridKey.of(tx, north - 1)]
+              if bs and bs.authored and bs.class == s.class then
+                north = north - 1
+              else
+                break
+              end
+            end
+            while front - ty < 6 do
+              local bs = S.shapeAt[GridKey.of(tx, front + 1)]
+              if bs and bs.authored and bs.class == s.class then
+                front = front + 1
+              else
+                break
+              end
+            end
+            local row = math.min(ty, front - math.floor(h / 8))
+            if row < north then
+              -- the whole run folded onto the face: top with the drawn
+              -- row just above it when that row is furniture too (a
+              -- bookcase wearing its shelf-top trim), else with the
+              -- run's own top row
+              local above = S.shapeAt[GridKey.of(tx, north - 1)]
+              row = (above and above.authored and above.art == "upright")
+                    and (north - 1) or north
+            end
+            topTile = S.tileAt[GridKey.of(tx, row)]
+          end
+          -- water's surface, and only water's: the recessed sheet itself,
+          -- never the ground's shoreline bands around it. A cell an object
+          -- stands on took the branch above and paints synthesized GROUND,
+          -- which is right -- a sign at the waterline stands on a plot, not
+          -- on the pond.
+          topQuad(x0, z0, h, topTile,
+                  s.art == "upright" and VOLUME_TOP_SHADE or 1,
+                  (s.class == "water") and waterPush or nil)
+        end
+
+        -- sides: 8px bands wherever the neighbour is lower. Band k spans
+        -- heights [8k, 8k+8) and shows one full tile of art; a partial
+        -- band crops the art rows to match, so nothing ever stretches.
+        for _, side in ipairs(SIDES) do
+          local nh = heightAt(tx + side[1], ty + side[2])
+          if nh < h then
+            local d = side[3]
+            -- the columns flanking this face, for the inside-corner term:
+            -- fixed for the whole face, so they are read once rather than
+            -- once per 8px band
+            local lat = LATERAL[d]
+            local hl = lat and heightAt(tx + lat[1], ty + lat[2]) or 0
+            local hr = lat and heightAt(tx + lat[3], ty + lat[4]) or 0
+            for band = math.floor(nh / 8), math.ceil(h / 8) - 1 do
+              local y0 = math.max(nh, band * 8)
+              local y1 = math.min(h, band * 8 + 8)
+              if y1 > y0 then
+                local src, shade = tile, Voxel3D.FACE_SHADE[d]
+                if run then
+                  -- fold the structure's artwork up this face: band k
+                  -- samples the map row k tiles north of the structure's
+                  -- front, clamped to its extent. The south face is the
+                  -- drawing itself (full brightness); the other sides wear
+                  -- the same rows darkened, so a building's flank matches
+                  -- its face instead of smearing one tile
+                  if d == 6 then
+                    src = map:tileAt(tx, math.min(run.front,
+                                                  run.north + band))
+                  else
+                    src = map:tileAt(tx, math.max(run.north,
+                                                  run.front - band))
+                  end
+                  if d == 5 then shade = 1 end
+                elseif s.art == "upright" then
+                  -- profile-authored upright (a pinned wall or furniture
+                  -- box): fold the drawing up the face, band 0 the
+                  -- structure's southmost same-class row and higher bands
+                  -- the rows north of it, repeating past the top.  The
+                  -- south face is the drawing itself (full brightness);
+                  -- flanks and back wear the same front stack darkened, so
+                  -- a desk's side matches its face instead of smearing a
+                  -- different jumble per row.
+                  if d == 5 then shade = 1 end
+                  local front = ty
+                  while front < ty + 6 do
+                    local fs2 = S.shapeAt[GridKey.of(tx, front + 1)]
+                    if fs2 and fs2.authored and fs2.class == s.class then
+                      front = front + 1
+                    else
+                      break
+                    end
+                  end
+                  local fk = GridKey.of(tx, front - band)
+                  local fs = S.shapeAt[fk]
+                  if fs and fs.authored and fs.class == s.class then
+                    src = S.tileAt[fk]
+                  end
+                end
+                sideQuad(d, x0, z0, y0, y1, src,
+                         (band * 8 + 8) - y1, (band * 8 + 8) - y0,
+                         sideShades(hl, hr, y0, y1, y0 <= nh, shade))
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+
+  -- Prebuilt quads from Structures (per-pixel voxel props, lathed
+  -- columns) plus the round-tree stamps expanded in place. Keep rules,
+  -- by the quad's own extent:
+  --   body-only   the quad must overlap the OPEN body interval -- a
+  --               neighbour's ring props must not march past its edge
+  --               into this map, and a quad lying exactly ON the edge
+  --               plane would z-fight the map that owns that plane.
+  --   full        anything overlapping the body stays whole (props that
+  --               straddle the edge no longer shed their outer half);
+  --               pure ring quads drop when they touch a neighbour body
+  --               (maskedClosed), which is what strings of seam pixels
+  --               were: fragments of dropped border trees whose centers
+  --               sat exactly on the boundary line.
+  local bw, bh = tw * 8, th * 8
+  local function keepQuad(x0, z0, x1, z1)
+    local overBody = x1 > 0 and x0 < bw and z1 > 0 and z0 < bh
+    if bodyOnly then return overBody end
+    return overBody or not maskedClosed(x0, z0, x1, z1)
+  end
+
+  -- A face lying EXACTLY on a body boundary plane is ambiguous to the
+  -- rect tests above: a body structure's outward facade (a Saffron row
+  -- house whose front row is the map's last row, its south wall on the
+  -- shared plane with Route 6) and the inward face of a ring scrap
+  -- occupy the same degenerate rect, and the strict overBody plus the
+  -- closed mask dropped BOTH -- which is why those facades were missing.
+  -- The winding tells them apart: a face pointing AWAY from the body
+  -- belongs to this map's own edge-row structure and nothing in the
+  -- neighbour will ever draw that plane, so it stays; a face pointing
+  -- INTO the body is the scrap the mask rules exist to kill, and falls
+  -- through to them.
+  local function outwardOnEdge(q, x0, z0, x1, z1)
+    if z0 == z1 and (z0 == 0 or z0 == bh) and x1 > 0 and x0 < bw then
+      local nz = (q[2][1] - q[1][1]) * (q[3][2] - q[1][2])
+                 - (q[2][2] - q[1][2]) * (q[3][1] - q[1][1])
+      return (z0 == bh and nz > 0) or (z0 == 0 and nz < 0)
+    end
+    if x0 == x1 and (x0 == 0 or x0 == bw) and z1 > 0 and z0 < bh then
+      local nx = (q[2][2] - q[1][2]) * (q[3][3] - q[1][3])
+                 - (q[2][3] - q[1][3]) * (q[3][2] - q[1][2])
+      return (x0 == bw and nx > 0) or (x0 == 0 and nx < 0)
+    end
+    return false
+  end
+
+  local scUV = { { 0, 0 }, { 0, 0 }, { 0, 0 }, { 0, 0 } }
+  local scObject = { { 0, 0, 0 }, { 0, 0, 0 },
+                     { 0, 0, 0 }, { 0, 0, 0 } }
+  local function quadUV(q)
+    if q.uv then return q.uv end
+    for i = 1, 4 do
+      scUV[i][1], scUV[i][2] = q.u, q.v
+    end
+    return scUV
+  end
+
+  for _, q in ipairs(S.objectQuads) do
+    Budget.tick()
+    -- Building stamps retain their immutable template and placement offset;
+    -- materialize into reusable corners before bounds, culling, and push.
+    local source, drawQ = q, q
+    if q.localQ then
+      source = q.localQ
+      local ox, oz = q.offsetX, q.offsetZ
+      for i = 1, 4 do
+        local c, d = source[i], scObject[i]
+        d[1], d[2], d[3] = c[1] + ox, c[2], c[3] + oz
+      end
+      drawQ = scObject
+    end
+    local x0 = math.min(drawQ[1][1], drawQ[2][1], drawQ[3][1], drawQ[4][1])
+    local x1 = math.max(drawQ[1][1], drawQ[2][1], drawQ[3][1], drawQ[4][1])
+    local z0 = math.min(drawQ[1][3], drawQ[2][3], drawQ[3][3], drawQ[4][3])
+    local z1 = math.max(drawQ[1][3], drawQ[2][3], drawQ[3][3], drawQ[4][3])
+    -- q.own: a body-anchored structure's own quad (a building placed by
+    -- Buildings.build, whose scan never leaves the body). Exempt from
+    -- the edge keep-rules entirely: its eave legitimately overhangs the
+    -- boundary plane into the neighbour's airspace, and no variant of
+    -- the neighbour will ever draw that geometry
+    if q.own or outwardOnEdge(drawQ, x0, z0, x1, z1)
+       or keepQuad(x0, z0, x1, z1) then
+      push({ drawQ[1], drawQ[2], drawQ[3], drawQ[4] },
+           quadUV(q), groundShades(q, q.shade))
+    end
+  end
+
+  -- round-tree stamps: the shared hull template translated per cell,
+  -- through reusable scratch corners so expansion allocates nothing.
+  -- A hull spans at most its own footprint -- one 16px cell unless the
+  -- stamp carries a wider radius (the 2x2-cell canopy groups) -- so one
+  -- rect test usually answers for the whole stamp: strictly interior
+  -- stamps keep every quad, ring stamps buried under a neighbour body
+  -- (or, body-only, ring stamps full stop) skip without touching their
+  -- quads.
+  --
+  -- A stamp is ONE tree, and the tree is atomic: the mask must never
+  -- cull its quads piecemeal. A stamp straddling the transition line
+  -- (partly under a neighbour body, partly on this map's ring) used to
+  -- fall through to per-quad keepQuad, which dropped every quad touching
+  -- the mask -- leaving a tree cut in half along the seam. The trunk
+  -- (the stamp centre) decides instead: a trunk under a neighbour body
+  -- is a tree that would rise through the neighbour's flat ground, so
+  -- the whole stamp goes; a trunk on this map's side keeps the whole
+  -- stamp, canopy overhang and all -- the overhang is above the
+  -- neighbour's ground, which is what a tree at a road edge does, and
+  -- the depth buffer sorts it against the neighbour's own geometry.
+  local sc = { { 0, 0, 0 }, { 0, 0, 0 }, { 0, 0, 0 }, { 0, 0, 0 } }
+  for _, st in ipairs(S.roundStamps or {}) do
+    local mx, mz = st.mx, st.mz
+    local sr = st.r or 8
+    local sx0, sz0, sx1, sz1 = mx - sr, mz - sr, mx + sr, mz + sr
+    local skipAll
+    if bodyOnly then
+      -- A neighbour contributes only its body. Same atomic rule as the
+      -- full variant, against the body rect instead of the mask rects:
+      -- the trunk (stamp centre) decides, so a body tree whose canopy
+      -- overhangs the body edge keeps every quad (a tree at the forest
+      -- edge legitimately overhangs the seam), and a ring tree whose
+      -- canopy pokes INTO the body is dropped whole -- its half-drawn
+      -- canopy used to float over the neighbour's ground with no trunk
+      -- (the per-quad slice), which read as a tree cut in half.
+      local centreInBody = mx >= 0 and mx <= bw and mz >= 0 and mz <= bh
+      skipAll = not centreInBody
+    else
+      -- Full variant: the mask rects are where connected neighbour
+      -- BODIES sit. A trunk under one is a tree that would rise through
+      -- the neighbour's flat ground -- the whole stamp goes. A trunk on
+      -- this map's side keeps the whole stamp, canopy overhang and all:
+      -- the overhang is above the neighbour's ground, which is what a
+      -- tree at a road edge does, and the depth buffer sorts it against
+      -- the neighbour's own geometry.
+      local centreInMask = false
+      if masks then
+        for _, mk in ipairs(masks) do
+          if mx >= mk[1] and mx <= mk[3] and mz >= mk[2] and mz <= mk[4] then
+            centreInMask = true
+            break
+          end
+        end
+      end
+      skipAll = centreInMask
+    end
+    if not skipAll then
+      for _, q in ipairs(st.quads) do
+        Budget.tick()
+        for i = 1, 4 do
+          local c, s2 = q[i], sc[i]
+          s2[1] = c[1] + mx
+          s2[2] = c[2]
+          s2[3] = c[3] + mz
+        end
+        push(sc, quadUV(q), groundShades(sc, q.shade))
+      end
+    end
+  end
+end
+
+return GeometryBuilder
